cursor-marketplace/netbox-mcp/pyproject.toml
[build-system]
requires = ["uv_build>=0.10.4,<0.11.0"]
build-backend = "uv_build"

[project]
name = "netbox-mcp"
version = "2.10.5"
description = "A read-only MCP server for NetBox"
readme = "README.md"
requires-python = ">=3.12"
license = "Apache-2.0"
keywords = ["mcp", "model-context-protocol", "netbox", "infrastructure"]
dependencies = [
    "fastmcp>=3.0.0,<4",
    "mcp-common>=0.11.0",
    "requests>=2.31.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "typer>=0.15.0",
]

[tool.uv.sources]
# mcp-common 0.11.x with testing.eval scorers (issues #42-#45)
mcp-common = { git = "https://github.com/vhspace/mcp-common", rev = "b504d774850e04991640416f79b4d31a32a5db0c" }

[project.scripts]
netbox-mcp = "netbox_mcp.server:main"
netbox-cli = "netbox_mcp.cli:main"

[project.optional-dependencies]
eval = [
    "inspect-ai",
    "openai>=1.0",
    "tenacity>=8.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "anyio>=4.0",
    "pytest-httpx>=0.34",
    "inline-snapshot>=0.15",
    "mypy>=1.13",
    "types-requests>=2.31",
    "ruff>=0.8",
    "pre-commit>=4.0",
    "pytest-cov>=7.0.0",
    "python-semantic-release>=10.4.1",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]
ignore = ["E402", "E501", "SIM102", "SIM105", "SIM108", "SIM117", "UP047"]

[tool.ruff.lint.per-file-ignores]
"src/netbox_mcp/cli.py" = ["B008"]  # typer.Option() in function defaults is intentional

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: unit tests",
    "integration: integration tests",
    "e2e: end-to-end tests",
]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
version_variables = [
    "src/netbox_mcp/__init__.py:__version__",
    ".cursor-plugin/plugin.json:\"version\"",
    ".claude-plugin/plugin.json:\"version\"",
]
branch = "main"
upload_to_vcs_release = true
tag_format = "v{version}"

[tool.semantic_release.commit_parser_options]
allowed_tags = ["feat", "fix", "docs", "chore", "refactor", "test", "ci", "perf"]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]

[tool.semantic_release.changelog]
changelog_file = "CHANGELOG.md"
